fix: Async system tests were not unwrapping async_generators (#1086)
Currently the system tests were failing with a message "TypeError:
'async_generator' object is not callable".

This is because newer versions of pytest_asyncio uses "strict" mode
by default which does not automatically convert fixtures to async
fixtures in async test cases. This became an issue as of pytest_asyncio
version 19.

The solution is to either switch the mode to "auto" in the nox
commandline or to mark the async fixtures as async fixtures.

I've opted for the second to avoid making this hard to discover.