fix(auth): Add temporary patch to workload cert logic to accomodate Cloud Run mis-configuration (#1880)
This patch adds a fallback logic to look for Cloud Run cert/keys in the
well-known location if the cert config contains the exact incorrect
cert/key paths AND the incorrect cert/key paths point to non-existent
files.

Note: This patch will be reverted sometime in Jan 2026, after Cloud Run
environment is updated with the correct cert configs. The revert will be
tracked by #1881