feat: Recognize workload certificate config in has_default_client_cert_source for mTLS for Agentic Identities (#1907)
feat: Recognize workload certificate config in
`mTLS.has_default_client_cert_source` for mTLS for Agentic Identities.
This is done as part of a fix for auto-enablement of mTLS, which depends
on `mTLS.has_default_client_cert_source`. This method returned `False`
because it only checked for gcloud cert configs and secure connect
metadata, but not the workload-provided certificate_config.json. Thus,
this change is necessary to help enable mTLS for Agentic Identities on
GKE and Cloud Run Workloads.

---------

Signed-off-by: Radhika Agrawal <agrawalradhika@google.com>